rename attention_cnn->cnn, feature_fusion->multimodal_fusion
