+++ kale/embed/cnn.py
