+++ tests/embed/test_attention_cnn.py
@@ -4,7 +4,7 @@ import torch
 import torch.nn as nn
 
 from kale.embed.attention import BANLayer
-from kale.embed.attention_cnn import CNNTransformer
+from kale.embed.cnn import CNNTransformer
 from kale.prepdata.tensor_reshape import seq_to_spatial
 from kale.utils.seed import set_seed
